Completed ~next_power command
Finished helper functions for ~next_power
Split destiny api helper functions into their own cog.
Lots and lots of other changes
