+++ cogs/error_handling_cogs.py
@@ -40,6 +40,7 @@ class error_handling_cogs(commands.Cog):
         if isinstance(error, errors.Error):
             await ctx.message.author.create_dm()
             await ctx.message.author.dm_channel.send(f'{str(error)}')
+            print(f'Custom error: {str(error)}')
 
         #because we only have role checks we know if the checks fail it was a role error
         elif isinstance(error, commands.errors.CheckFailure):
@@ -72,7 +73,12 @@ class error_handling_cogs(commands.Cog):
 
             #send error message to server admin
             await admin.create_dm()
-            await admin.dm_channel.send(f'Command error occured at {now}\nUser: {ctx.message.author.name}\nMessage: {ctx.message.content}\nTraceback: {traceback.format_exc()}\nError: {error}')
+            await admin.dm_channel.send(f'Command error occured at {now}\nUser: {ctx.message.author.name}\nMessage: {ctx.message.content}\nTraceback: {traceback.format_exc()}\nError: ' + '{}: {}'.format(type(error).__name__, error))
+
+            #delete message that caused error to keep channels clean
+            await ctx.message.delete()
+
+            raise error
 
         #check to see if they user was trying to cross out a message and accidentally triggered the bot, if not, delete their message
         if(ctx.message.content.split()[0][1] != "~"):
